backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
urllib3==2.5.0
